+++ packages/react/src/atomic/Button/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The Button component allows us to initiate actions, make state or page changes.
 
+## 2023-02-10 `2.2.0`
+
+### 🎉 New features
+
+- Added polymorphic typing support to component. ([#89](https://github.com/TiendaNube/nimbus-design-system/pull/89) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `as` property to the Component. ([#89](https://github.com/TiendaNube/nimbus-design-system/pull/89) by [@juniorconquista](https://github.com/juniorconquista))
+
+### 💡 Others
+
+- Removed subcomponent `Button.Link`. [#89](https://github.com/TiendaNube/nimbus-design-system/pull/#89) by [@juniorconquista](https://github.com/juniorconquista))
+
 ## 2023-02-06 `2.1.0`
 
 ### 🎉 New features